requirements.txt
pillow
mysql-connector-python
requests
tk
